Continue refactoring the subscription mechanism
