+++ include/abb_librws/rws_poco_client.h
@@ -37,293 +37,196 @@
 #ifndef RWS_POCO_CLIENT_H
 #define RWS_POCO_CLIENT_H
 
-#include "Poco/Mutex.h"
-#include "Poco/Net/HTTPClientSession.h"
-#include "Poco/Net/HTTPCredentials.h"
-#include "Poco/Net/HTTPResponse.h"
-#include "Poco/Net/WebSocket.h"
-#include "Poco/SharedPtr.h"
+#include <abb_librws/rws_poco_result.h>
+
+#include <Poco/Mutex.h>
+#include <Poco/Net/HTTPClientSession.h>
+#include <Poco/Net/HTTPCredentials.h>
+#include <Poco/Net/HTTPResponse.h>
+#include <Poco/Net/WebSocket.h>
+
+#include <memory>
 
 namespace abb
 {
 namespace rws
 {
   /**
-   * \brief A struct for containing the result of a communication.
+   * \brief A class for a simple client based on POCO.
    */
-  struct POCOResult
+  class POCOClient
   {
+  public:
     /**
-     * \brief An enum for specifying a general status.
-     */
-    enum GeneralStatus
-    {
-      UNKNOWN,                         ///< Unknown status.
-      OK,                              ///< Ok status.
-      WEBSOCKET_NOT_ALLOCATED,         ///< The WebSocket has not been allocated.
-      EXCEPTION_POCO_INVALID_ARGUMENT, ///< POCO invalid argument exception
-      EXCEPTION_POCO_TIMEOUT,          ///< POCO timeout exception.
-      EXCEPTION_POCO_NET,              ///< POCO net exception.
-      EXCEPTION_POCO_WEBSOCKET         ///< POCO WebSocket exception.
-    };
-
-    /**
-     * \brief A struct for containing POCO info.
-     */
-    struct POCOInfo
-    {
-      /**
-       * \brief A struct for containing HTTP info.
-       */
-      struct HTTPInfo
-      {
-        /**
-         * \brief A struct for containing HTTP request info.
-         */
-        struct RequestInfo
-        {
-          /**
-           * \brief Method used for the request.
-           */
-          std::string method;
-
-          /**
-           * \brief URI used for the request.
-           */
-          std::string uri;
-
-          /**
-           * \brief Content used for the request.
-           */
-          std::string content;
-        };
-
-        /**
-         * \brief A struct for containing HTTP response info.
-         */
-        struct ResponseInfo
-        {
-          /**
-           * \brief Response status.
-           */
-          Poco::Net::HTTPResponse::HTTPStatus status;
-
-          /**
-           * \brief Response header info.
-           */
-          std::string header_info;
-
-          /**
-           * \brief Response content.
-           */
-          std::string content;
-
-          /**
-           * \brief A default constructor.
-           */
-          ResponseInfo() : status(Poco::Net::HTTPResponse::HTTP_OK) {}
-        };
-
-        /**
-         * \brief Info about a HTTP request.
-         */
-        RequestInfo request;
-
-         /**
-          * \brief Info about a HTTP response.
-          */
-        ResponseInfo response;
-      };
-
-      /**
-       * \brief A struct for containing WebSocket info.
-       */
-      struct WebSocketInfo
-      {
-        /**
-         * \brief Flags from a received WebSocket frame.
-         */
-        int flags;
-
-        /**
-         * \brief Content from a received WebSocket frame.
-         */
-        std::string frame_content;
-
-        /**
-         * \brief A default constructor.
-         */
-        WebSocketInfo() : flags(0) {};
-      };
-
-      /**
-       * \brief Container for HTTP info.
-       */
-      HTTPInfo http;
-
-      /**
-       * \brief Container for WebSocket info.
-       */
-      WebSocketInfo websocket;
-    };
-
-    /**
-     * \brief Container for a general status.
-     */
-    GeneralStatus status;
-
-    /**
-     * \brief Container for an exception message (if one occurred).
-     */
-    std::string exception_message;
-
-    /**
-     * \brief Container for POCO info.
-     */
-    POCOInfo poco_info;
-
-    /**
-     * \brief A default constructor.
-     */
-    POCOResult() : status(UNKNOWN) {};
-
-    /**
-     * \brief A method for adding info from a HTTP request.
+     * \brief A constructor.
      *
+     * \param ip_address for the remote server's IP address.
+     * \param port for the remote server's port.
+     * \param username for the username to the remote server's authentication process.
+     * \param password for the password to the remote server's authentication process.
+     */
+    POCOClient(const std::string& ip_address,
+              const Poco::UInt16 port,
+              const std::string& username,
+              const std::string& password)
+    :
+    http_client_session_(ip_address, port),
+    http_credentials_(username, password)
+    {
+      http_client_session_.setKeepAlive(true);
+      http_client_session_.setTimeout(Poco::Timespan(DEFAULT_HTTP_TIMEOUT));
+    }
+
+    /**
+     * \brief A destructor.
+     */
+    ~POCOClient() {}
+
+    /**
+     * \brief A method for sending a HTTP GET request.
+     *
+     * \param uri for the URI (path and query).
+     *
+     * \return POCOResult containing the result.
+     */
+    POCOResult httpGet(const std::string& uri);
+
+    /**
+     * \brief A method for sending a HTTP POST request.
+     *
+     * \param uri for the URI (path and query).
+     * \param content for the request's content.
+     *
+     * \return POCOResult containing the result.
+     */
+    POCOResult httpPost(const std::string& uri, const std::string& content = "");
+
+    /**
+     * \brief A method for sending a HTTP PUT request.
+     *
+     * \param uri for the URI (path and query).
+     * \param content for the request's content.
+     *
+     * \return POCOResult containing the result.
+     */
+    POCOResult httpPut(const std::string& uri, const std::string& content = "");
+
+    /**
+     * \brief A method for sending a HTTP DELETE request.
+     *
+     * \param uri for the URI (path and query).
+     *
+     * \return POCOResult containing the result.
+     */
+    POCOResult httpDelete(const std::string& uri);
+
+    /**
+     * \brief A method for setting the HTTP communication timeout.
+     *
+     * \note This method resets the internal HTTP client session, causing the
+     *       RWS server (robot controller) to send a new cookie. The RWS
+     *       session id is not changed.
+     *
+     * \param timeout for the HTTP communication timeout [microseconds].
+     */
+    void setHTTPTimeout(const Poco::Int64 timeout)
+    {
+      http_client_session_.setTimeout(Poco::Timespan(timeout));
+      http_client_session_.reset();
+    }
+
+
+    /**
+     * \brief A method for connecting a WebSocket.
+     *
+     * \param uri for the URI (path and query).
+     * \param protocol for the WebSocket protocol.
+     * \param timeout for the WebSocket communication timeout [microseconds].
+     *
+     * \return Pointer to a client WebSocket.
+     * 
+     * \throw \a std::runtime_error if something goes wrong
+     */
+    std::unique_ptr<Poco::Net::WebSocket> webSocketConnect(const std::string& uri, const std::string& protocol, const Poco::Int64 timeout);
+
+
+  private:
+    /**
+     * \brief A method for making a HTTP request.
+     *
+     * \param method for the request's method.
+     * \param uri for the URI (path and query).
+     * \param content for the request's content.
+     *
+     * \return POCOResult containing the result.
+     */
+    POCOResult makeHTTPRequest(const std::string& method,
+                              const std::string& uri = "/",
+                              const std::string& content = "");
+
+    /**
+     * \brief A method for sending and receiving HTTP messages.
+     *
+     * \param result for the result.
      * \param request for the HTTP request.
-     * \param request_content for the HTTP request's content.
-     */
-    void addHTTPRequestInfo(const Poco::Net::HTTPRequest& request, const std::string& request_content = "");
-
-    /**
-     * \brief A method for adding info from a HTTP response.
-     *
      * \param response for the HTTP response.
-     * \param response_content for the HTTP response's content.
+     * \param request_content for the request's content.
      */
-    void addHTTPResponseInfo(const Poco::Net::HTTPResponse& response, const std::string& response_content = "");
+    void sendAndReceive(POCOResult& result,
+                        Poco::Net::HTTPRequest& request,
+                        Poco::Net::HTTPResponse& response,
+                        const std::string& request_content);
 
     /**
-     * \brief A method for adding info from a received WebSocket frame.
+     * \brief A method for performing authentication.
      *
-     * \param flags for the received WebSocket frame's flags.
-     * \param frame_content for the received WebSocket frame's content.
+     * \param result for the result.
+     * \param request for the HTTP request.
+     * \param response for the HTTP response.
+     * \param request_content for the request's content.
      */
-    void addWebSocketFrameInfo(const int flags, const std::string& frame_content);
+    void authenticate(POCOResult& result,
+                      Poco::Net::HTTPRequest& request,
+                      Poco::Net::HTTPResponse& response,
+                      const std::string& request_content);
 
     /**
-     * \brief A method to map the general status to a std::string.
+     * \brief A method for extracting and storing information from a cookie string.
      *
-     * \return std::string containing the mapped general status.
+     * \param cookie_string for the cookie string.
      */
-    std::string mapGeneralStatus() const;
+    void extractAndStoreCookie(const std::string& cookie_string);
 
     /**
-     * \brief A method to map the opcode of a received WebSocket frame.
-     *
-     * \return std::string containing the mapped opcode.
+     * \brief Static constant for the default HTTP communication timeout [microseconds].
      */
-    std::string mapWebSocketOpcode() const;
+    static const Poco::Int64 DEFAULT_HTTP_TIMEOUT = 400e3;
+
 
     /**
-     * \brief A method to construct a text representation of the result.
-     *
-     * \param verbose indicating if the log text should be verbose or not.
-     * \param indent for indentation.
-     *
-     * \return std::string containing the text representation.
+     * \brief A mutex for protecting the clients's HTTP resources.
      */
-    std::string toString(const bool verbose = false, const size_t indent = 0) const;
+    Poco::Mutex http_mutex_;
+
+    /**
+     * \brief A HTTP client session.
+     */
+    Poco::Net::HTTPClientSession http_client_session_;
+
+    /**
+     * \brief HTTP credentials for the remote server's access authentication process.
+     */
+    Poco::Net::HTTPCredentials http_credentials_;
+
+    /**
+     * \brief A container for cookies received from a server.
+     */
+    Poco::Net::NameValueCollection cookies_;
   };
 
 
-/**
- * \brief A class for a simple client based on POCO.
- */
-class POCOClient
-{
-public:
   /**
-   * \brief A constructor.
-   *
-   * \param ip_address for the remote server's IP address.
-   * \param port for the remote server's port.
-   * \param username for the username to the remote server's authentication process.
-   * \param password for the password to the remote server's authentication process.
-   */
-  POCOClient(const std::string& ip_address,
-             const Poco::UInt16 port,
-             const std::string& username,
-             const std::string& password)
-  :
-  http_client_session_(ip_address, port),
-  http_credentials_(username, password)
-  {
-    http_client_session_.setKeepAlive(true);
-    http_client_session_.setTimeout(Poco::Timespan(DEFAULT_HTTP_TIMEOUT));
-  }
-
-  /**
-   * \brief A destructor.
-   */
-  ~POCOClient() {}
-
-  /**
-   * \brief A method for sending a HTTP GET request.
-   *
-   * \param uri for the URI (path and query).
-   *
-   * \return POCOResult containing the result.
-   */
-  POCOResult httpGet(const std::string& uri);
-
-  /**
-   * \brief A method for sending a HTTP POST request.
-   *
-   * \param uri for the URI (path and query).
-   * \param content for the request's content.
-   *
-   * \return POCOResult containing the result.
-   */
-  POCOResult httpPost(const std::string& uri, const std::string& content = "");
-
-  /**
-   * \brief A method for sending a HTTP PUT request.
-   *
-   * \param uri for the URI (path and query).
-   * \param content for the request's content.
-   *
-   * \return POCOResult containing the result.
-   */
-  POCOResult httpPut(const std::string& uri, const std::string& content = "");
-
-  /**
-   * \brief A method for sending a HTTP DELETE request.
-   *
-   * \param uri for the URI (path and query).
-   *
-   * \return POCOResult containing the result.
-   */
-  POCOResult httpDelete(const std::string& uri);
-
-  /**
-   * \brief A method for setting the HTTP communication timeout.
-   *
-   * \note This method resets the internal HTTP client session, causing the
-   *       RWS server (robot controller) to send a new cookie. The RWS
-   *       session id is not changed.
-   *
-   * \param timeout for the HTTP communication timeout [microseconds].
-   */
-  void setHTTPTimeout(const Poco::Int64 timeout)
-  {
-    http_client_session_.setTimeout(Poco::Timespan(timeout));
-    http_client_session_.reset();
-  }
-
-  
-  /**
-   * \brief A method for retrieving a substring in a string.
+   * \brief A function for retrieving a substring in a string.
    *
    * \param whole_string for the string containing the substring.
    * \param substring_start start of the substring.
@@ -334,85 +237,6 @@ public:
   std::string findSubstringContent(const std::string& whole_string,
                                    const std::string& substring_start,
                                    const std::string& substring_end);
-
-private:
-  /**
-   * \brief A method for making a HTTP request.
-   *
-   * \param method for the request's method.
-   * \param uri for the URI (path and query).
-   * \param content for the request's content.
-   *
-   * \return POCOResult containing the result.
-   */
-  POCOResult makeHTTPRequest(const std::string& method,
-                             const std::string& uri = "/",
-                             const std::string& content = "");
-
-  /**
-   * \brief A method for sending and receiving HTTP messages.
-   *
-   * \param result for the result.
-   * \param request for the HTTP request.
-   * \param response for the HTTP response.
-   * \param request_content for the request's content.
-   */
-  void sendAndReceive(POCOResult& result,
-                      Poco::Net::HTTPRequest& request,
-                      Poco::Net::HTTPResponse& response,
-                      const std::string& request_content);
-
-  /**
-   * \brief A method for performing authentication.
-   *
-   * \param result for the result.
-   * \param request for the HTTP request.
-   * \param response for the HTTP response.
-   * \param request_content for the request's content.
-   */
-  void authenticate(POCOResult& result,
-                    Poco::Net::HTTPRequest& request,
-                    Poco::Net::HTTPResponse& response,
-                    const std::string& request_content);
-
-  /**
-   * \brief A method for extracting and storing information from a cookie string.
-   *
-   * \param cookie_string for the cookie string.
-   */
-  void extractAndStoreCookie(const std::string& cookie_string);
-
-  /**
-   * \brief Static constant for the default HTTP communication timeout [microseconds].
-   */
-  static const Poco::Int64 DEFAULT_HTTP_TIMEOUT = 400e3;
-
-  /**
-   * \brief Static constant for the socket's buffer size.
-   */
-  static const size_t BUFFER_SIZE = 1024;
-
-  /**
-   * \brief A mutex for protecting the clients's HTTP resources.
-   */
-  Poco::Mutex http_mutex_;
-
-  /**
-   * \brief A HTTP client session.
-   */
-  Poco::Net::HTTPClientSession http_client_session_;
-
-  /**
-   * \brief HTTP credentials for the remote server's access authentication process.
-   */
-  Poco::Net::HTTPCredentials http_credentials_;
-
-  /**
-   * \brief A container for cookies received from a server.
-   */
-  Poco::Net::NameValueCollection cookies_;
-};
-
 } // end namespace rws
 } // end namespace abb
