+++ include/abb_librws/rws_websocket_info.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <string>
+
+
+namespace abb :: rws
+{
+    /**
+     * \brief A struct for containing WebSocket info.
+     */
+    struct WebSocketInfo
+    {
+        /**
+         * \brief Flags from a received WebSocket frame.
+         */
+        int flags = 0;
+
+        /**
+         * \brief Content from a received WebSocket frame.
+         */
+        std::string frame_content;
+    };
+}
